% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.CopyNumber.Genes.All.R
\name{Analyses.CopyNumber.Genes.All}
\alias{Analyses.CopyNumber.Genes.All}
\title{Retrieve all data by genes Gistic2 results.}
\usage{
Analyses.CopyNumber.Genes.All(format = "json", cohort = "", gene = "",
  tcga_participant_barcode = "", page = "1", page_size = "250",
  sort_by = "cohort")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{cohort}{Narrow search to one or more TCGA disease cohorts from the scrollable list. Multiple values are allowed ACC,BLCA,BRCA,CESC,CHOL,COAD,COADREAD,DLBC,ESCA,FPPP,GBM,GBMLGG,HNSC,KICH,KIPAN,KIRC,KIRP,LAML,LGG,LIHC,LUAD,LUSC,MESO,OV,PAAD,PCPG,PRAD,READ,SARC,SKCM,STAD,STES,TGCT,THCA,THYM,UCEC,UCS,UVM.}

\item{gene}{Comma separated list of gene name(s). Multiple values are allowed .}

\item{tcga_participant_barcode}{Comma separated list of TCGA participant barcodes (e.g. TCGA-GF-A4EO). Multiple values are allowed .}

\item{page}{Which page (slice) of entire results set should be returned.  Multiple values are allowed . Default value is 1.}

\item{page_size}{Number of records per page of results.  Max is 2000. Multiple values are allowed . Default value is 250.}

\item{sort_by}{Which column in the results should be used for sorting paginated results? Default value is cohort. While tcga_participant_barcode,cohort,gene are available.}
}
\description{
This service provides access to the Gistic2 all_data_by_genes.txt output data. This data is a gene-level table of copy number values for all samples. The returned copy number values are in units (copy number - 2) so that no amplification or deletion is 0, genes with amplifications have positive values, and genes with deletions are negative values. The data are converted from marker level to gene level using the extreme method: a gene is assigned the greatest amplification or the least deletion value among the markers it covers. Results may be filtered by cohort, gene, or barcode, but at least one gene or barcode must be supplied.
}

