% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.CopyNumber.Genes.Deleted.R
\name{Analyses.CopyNumber.Genes.Deleted}
\alias{Analyses.CopyNumber.Genes.Deleted}
\title{Retrieve Gistic2 significantly deleted genes results.}
\usage{
Analyses.CopyNumber.Genes.Deleted(format = "json", cohort = "", gene = "",
  q = "", page = "1", page_size = "250", sort_by = "cohort")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{cohort}{Narrow search to one or more TCGA disease cohorts from the scrollable list. Multiple values are allowed ACC,BLCA,BRCA,CESC,CHOL,COAD,COADREAD,DLBC,ESCA,FPPP,GBM,GBMLGG,HNSC,KICH,KIPAN,KIRC,KIRP,LAML,LGG,LIHC,LUAD,LUSC,MESO,OV,PAAD,PCPG,PRAD,READ,SARC,SKCM,STAD,STES,TGCT,THCA,THYM,UCEC,UCS,UVM.}

\item{gene}{Comma separated list of gene name(s). Multiple values are allowed .}

\item{q}{Only return results with Q-value &lt;= given threshold.}

\item{page}{Which page (slice) of entire results set should be returned.  Multiple values are allowed . Default value is 1.}

\item{page_size}{Number of records per page of results.  Max is 2000. Multiple values are allowed . Default value is 250.}

\item{sort_by}{Which column in the results should be used for sorting paginated results? Default value is cohort. While q,cohort,gene are available.}
}
\description{
This service provides access to the Gistic2 del_genes.conf_99.txt output data.  At least 1 gene or cohort must be supplied.
}

