% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.CopyNumber.Genes.Thresholded.R
\name{Analyses.CopyNumber.Genes.Thresholded}
\alias{Analyses.CopyNumber.Genes.Thresholded}
\title{Retrieve all thresholded by genes Gistic2 results.}
\usage{
Analyses.CopyNumber.Genes.Thresholded(format = "csv", cohort = "",
  gene = "", tcga_participant_barcode = "", page = 1, page_size = 250,
  sort_by = "cohort")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}

\item{gene}{A character vector of gene symbols. At least one gene OR Barcode
is required.}

\item{tcga_participant_barcode}{A character vector containing TCGA Barcodes.
Empty string returns all patients. Either one gene OR barcode is required.}

\item{page}{Subset to be returned.}

\item{page_size}{Number of records per page, max. is 2000.}

\item{sort_by}{character indicating the to column which is used for sorting.}
}
\value{
A \code{list}, if format is \code{json}, elsewise a \code{data.frame}
}
\description{
This service provides access to the Gistic2 all_thresholded_by_genes.txt
output data. A gene-level table of discrete amplification and deletion
indicators for all samples. A table value of 0 means no amplification or
deletion above the threshold. Amplifications are positive numbers: 1 means
amplification above the amplification threshold; 2 means amplifications
larger to the arm level amplifications observed for the sample. Deletions are
represented by negative table values: -1 represents deletion beyond the
threshold; -2 means deletions greater than the minimum arm-level deletion
observed for the sample. Results maybe filtered by cohort, gene or barcode,
but at least one gene or barcode must be supplied.
}
\examples{
format = "csv"
cohort = "PRAD"
gene = "PTEN"
tcga_participant_barcode = c("TCGA-J4-A67O", "TCGA-G9-6496")
page = 1
page_size = 250
sort_by = "cohort"

obj = Analyses.CopyNumber.Genes.Thresholded(format = format,
                                cohort = cohort,
                                gene = gene,
                                tcga_participant_barcode = tcga_participant_barcode,
                                page = page,
                                page_size = page_size,
                                sort_by = sort_by)

gene = ""
obj = Analyses.CopyNumber.Genes.Thresholded(format = format,
                                cohort = cohort,
                                gene = gene,
                                tcga_participant_barcode = tcga_participant_barcode,
                                page = page,
                                page_size = page_size,
                                sort_by = sort_by)

format = "json"
obj = Analyses.CopyNumber.Genes.Thresholded(format = format,
                                cohort = cohort,
                                gene = gene,
                                tcga_participant_barcode = tcga_participant_barcode,
                                page = page,
                                page_size = page_size,
                                sort_by = sort_by)

}

