% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Analyses.FeatureTable.R
\name{Analyses.FeatureTable}
\alias{Analyses.FeatureTable}
\title{Retrieve aggregated analysis features table.}
\usage{
Analyses.FeatureTable(format = "csv", cohort = "", column = "",
  page = 1, page_size = 250)
}
\arguments{
\item{format}{Either \code{tsv} or \code{csv}, here \code{json} is not
possible, but comming soon.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}

\item{column}{Comma separated list of which data columns/fields to return.}

\item{page}{Subset to be returned.}

\item{page_size}{Number of records per page, max. is 2000.}
}
\value{
A \code{list}, if format is \code{json}, elsewise a \code{data.frame}
}
\description{
This service returns part or all of the so-called feature table; which
aggregates the most important findings across ALL pipelines in the GDAC
Firehose analysis workflow into a single table for simple access. For more
details please visit the online documentation (\url{https://confluence.broadinstitute.org/display/GDAC/Documentation/#Documentation-FeatureTable}).
Please note that the service is still undergoing experimental evaluation and
does not return JSON format.
}

