% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.Mutation.MAF.R
\name{Analyses.Mutation.MAF}
\alias{Analyses.Mutation.MAF}
\title{Retrieve MutSig final analysis MAF.}
\usage{
Analyses.Mutation.MAF(format = "json", cohort = "", tool = "MutSig2CV",
  gene = "", tcga_participant_barcode = "", column = "", page = "1",
  page_size = "250", sort_by = "cohort")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{cohort}{Narrow search to one or more TCGA disease cohorts from the scrollable list. Multiple values are allowed ACC,BLCA,BRCA,CESC,CHOL,COAD,COADREAD,DLBC,ESCA,FPPP,GBM,GBMLGG,HNSC,KICH,KIPAN,KIRC,KIRP,LAML,LGG,LIHC,LUAD,LUSC,MESO,OV,PAAD,PCPG,PRAD,READ,SARC,SKCM,STAD,STES,TGCT,THCA,THYM,UCEC,UCS,UVM.}

\item{tool}{Narrow search to include only data/results produced by the selected Firehose tool. Multiple values are allowed MutSig2.0,MutSig2CV. Default value is MutSig2CV.}

\item{gene}{Comma separated list of gene name(s). Multiple values are allowed .}

\item{tcga_participant_barcode}{Comma separated list of TCGA participant barcodes (e.g. TCGA-GF-A4EO). Multiple values are allowed .}

\item{column}{Comma separated list of which data columns/fields to return. Multiple values are allowed .}

\item{page}{Which page (slice) of entire results set should be returned.  Multiple values are allowed . Default value is 1.}

\item{page_size}{Number of records per page of results.  Max is 2000. Multiple values are allowed . Default value is 250.}

\item{sort_by}{Which column in the results should be used for sorting paginated results? Default value is cohort. While tcga_participant_barcode,cohort,tool,gene are available.}
}
\description{
This service returns columns from the MAF generated by MutSig. Results may be filtered by gene, cohort, tool, or barcode, but at least one gene OR barcode OR cohort must be given.  By default a subset consisting of the most commonly used columns will be returned, but that can be modified with the column parameter. Specifying 'all' in this parameter is a convenient way to return every column of the respective MAF, and has precedence over any any other column selection expression.  The complete list of column names that may be specified is <a href='http://firebrowse.org/api-docs/#!/Metadata/MAFColNames'>given here</a>.  For more information on the mutation data, and how it is processed by Firehose, please consult the <a href="https://confluence.broadinstitute.org/display/GDAC/Documentation#Documentation-MutationPipelines">pipeline documentation</a>.
}

