% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.Mutation.SMG.R
\name{Analyses.Mutation.SMG}
\alias{Analyses.Mutation.SMG}
\title{Retrieve Significantly Mutated Genes (SMG).}
\usage{
Analyses.Mutation.SMG(format = "json", cohort = "", tool = "MutSig2CV",
  rank = "", gene = "", q = "", page = "1", page_size = "250",
  sort_by = "rank")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{cohort}{Narrow search to one or more TCGA disease cohorts from the scrollable list. Multiple values are allowed ACC,BLCA,BRCA,CESC,CHOL,COAD,COADREAD,DLBC,ESCA,FPPP,GBM,GBMLGG,HNSC,KICH,KIPAN,KIRC,KIRP,LAML,LGG,LIHC,LUAD,LUSC,MESO,OV,PAAD,PCPG,PRAD,READ,SARC,SKCM,STAD,STES,TGCT,THCA,THYM,UCEC,UCS,UVM.}

\item{tool}{Narrow search to include only data/results produced by the selected Firehose tool. Multiple values are allowed MutSig2.0,MutSig2CV,MutSigCV. Default value is MutSig2CV.}

\item{rank}{Number of significant genes to return.}

\item{gene}{Comma separated list of gene name(s). Multiple values are allowed .}

\item{q}{Only return results with Q-value &lt;= given threshold.}

\item{page}{Which page (slice) of entire results set should be returned.  Multiple values are allowed . Default value is 1.}

\item{page_size}{Number of records per page of results.  Max is 2000. Multiple values are allowed . Default value is 250.}

\item{sort_by}{Which column in the results should be used for sorting paginated results? Default value is rank. While q,cohort,tool,gene,rank are available.}
}
\description{
This service provides a list of significantly mutated genes, as scored by MutSig.  It may be filtered by cohort, rank, gene, tool and/or Q-value threshold, but at least one cohort must be supplied.
}

