% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Querybuilder.R
\name{API.Version}
\alias{API.Version}
\title{Return the API version this package was build for}
\usage{
API.Version()
}
\value{
A character
}
\description{
Returns the API version this package was build for, should match with version
at \url{http://firebrowse.org/api-docs}.
}
\examples{
API.Version()

}

