% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Archives.StandardData.R
\name{Archives.StandardData}
\alias{Archives.StandardData}
\title{Retrieve standard data archives.}
\usage{
Archives.StandardData(format = "tsv", date = "", cohort = "",
  data_type = "", tool = "", platform = "", center = "", level = "",
  protocol = "", page = 1, page_size = 250, sort_by = "cohort")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{date}{Narrow search to one or more Firehose run date stamps. For a list
of available dates see \code{\link{Metadata.Dates}}}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}

\item{data_type}{Narrow search to one or more TCGA data types. Available data
types are provided by the data frame \code{data_types}.}

\item{tool}{Narrow search to include only data/results produced by the selected
Firehose tool.  Available tools are provided in the data frame \code{tools}.}

\item{platform}{A platforms name.}

\item{center}{A center name.}

\item{level}{choose data level from 1-4.}

\item{protocol}{Narrow search to one or more sample characterization
protocols, see \code{protocols} data.frame for available protocols.}

\item{page}{Subset to be returned.}

\item{page_size}{Number of records per page, max. is 2000.}

\item{sort_by}{character indicating the to column which is used for sorting.}
}
\value{
A \code{list}, if format is \code{json}, elsewise a \code{data.frame}
}
\description{
This service returns the archive URLs for our Firehose standard data runs,
providing a RESTful interface similar in spirit to the command line
firehose_get tool. The archives can be filtered based on date, cohort, data
type, platform, center, data level, and protocol.
}
\details{
\code{data_type}m \code{tool} and \code{protocol} are hard coded, since the
API provides no methods to query them. Therefor, and to keep always up to
date, you may consider looking at the website
\url{http://firebrowse.org/api-docs/} for all available data types.
}
\examples{
format = "tsv"
date = "2014_12_06"
cohort = "BRCA"
data_type = "CopyNumber"
level = 3
page = 1
page_size = 250

obj = Archives.StandardData(format = format,
                                date = date,
                                cohort = cohort,
                                data_type = data_type,
                                level = level,
                                page = page,
                                page_size = page_size)

cohort = "PRAD"
data_type = "MAF"
level = 3
obj = Archives.StandardData(format = format,
                                date = date,
                                cohort = cohort,
                                data_type = data_type,
                                level = level,
                                page = page,
                                page_size = page_size)
}

