% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/External_Data_Documentation.R
\docType{data}
\name{data_types}
\alias{data_types}
\title{data_types includes all available data types}
\format{A matrix}
\source{
\url{http://firebrowse.org/api-docs/#!/Archives/StandardData}
}
\usage{
data(data_types)
}
\description{
For two functions/queries (namely \code{archives_standard_data} and
\code{metadata_counts})a data_type, such as \code{miR} or \code{CopyNumber}
is required, but there is no API function listing all possible options. This
table indicates which function accepts which data type.
This table is hand curated, informations are directly gathered from source
listed below
}
\keyword{datasets}

