% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Metadata.Centers.R
\name{Metadata.Centers}
\alias{Metadata.Centers}
\title{Retrieve map of center name(s) to display name(s).}
\usage{
Metadata.Centers(format = "csv", center = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{center}{A center name.}
}
\description{
To return a map of all available centers to display names, exclude the center
parameter. Otherwise, choose only the centers you wish to be included in the
result.
}
\examples{
format = "csv"
center = c("intgen.org", "jhu-usc.edu")
obj = Metadata.Centers(format = format, center = center)
}

