% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.ClinicalNames.R
\name{Metadata.ClinicalNames}
\alias{Metadata.ClinicalNames}
\title{Retrieve names of all TCGA clinical data elements (CDEs).}
\usage{
Metadata.ClinicalNames(format = "csv")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}
}
\description{
Retrieve names of all patient-level clinical data elements (CDES) available
in TCGA, unioned across all disease cohorts. A CDE will be listed here only
when it has a value other than NA for at least 1 patient case in any disease
cohort. For more information on how these CDEs are processed see our pipeline
documentation.
\url{https://confluence.broadinstitute.org/display/GDAC/Documentation}
}
\examples{
format = "csv"
obj = Metadata.ClinicalNames(format = format)
}

