% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.ClinicalNames_FH.R
\name{Metadata.ClinicalNames_FH}
\alias{Metadata.ClinicalNames_FH}
\title{Retrieve names of CDEs normalized by Firehose and selected for analyses.}
\usage{
Metadata.ClinicalNames_FH(format = "json")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}
}
\description{
This service returns the names of patient-level clinical data elements (CDEs) that have been normalized by Firehose for use in analyses, unioned across all disease cohorts. For more information on how these CDEs are processed, see our <a href="https://confluence.broadinstitute.org/display/GDAC/Documentation">pipeline documentation</a>.
}

