% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.ClinicalTier1.R
\name{Metadata.ClinicalTier1}
\alias{Metadata.ClinicalTier1}
\title{Retrieve names of all available clinical data elements (CDEs).}
\usage{
Metadata.ClinicalTier1(format = "csv")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}
}
\description{
Retrieve names of all available tier 1 clinical data elements (CDEs), unioned
across all disease cohorts and patients. For more information on how these
CDEs are processed see our pipeline documentation
\url{https://confluence.broadinstitute.org/display/GDAC/Documentation}
}
\examples{
format = "csv"
obj = Metadata.ClinicalTier1(format = format)

format = "tsv"
obj = Metadata.ClinicalTier1(format = format)
}

