% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.Cohort.R
\name{Metadata.Cohort}
\alias{Metadata.Cohort}
\title{Retrieve formal disease name(s) for cohort abbreviation(s).}
\usage{
Metadata.Cohort(format = "json", cohort = "")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{cohort}{Narrow search to one or more TCGA disease cohorts from the scrollable list. Multiple values are allowed ACC,BLCA,BRCA,CESC,CHOL,COAD,COADREAD,DLBC,ESCA,FPPP,GBM,GBMLGG,HNSC,KICH,KIPAN,KIRC,KIRP,LAML,LGG,LIHC,LUAD,LUSC,MESO,OV,PAAD,PCPG,PRAD,READ,SARC,SKCM,STAD,STES,TGCT,THCA,THYM,UCEC,UCS,UVM.}
}
\description{
Retrieve formal disease name(s) for cohort abbreviation(s).
}

