% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.Cohorts.R
\name{Metadata.Cohorts}
\alias{Metadata.Cohorts}
\title{Retrieve map of cohort abbreviation(s) to cohort name(s).}
\usage{
Metadata.Cohorts(format = "csv", cohort = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}
}
\description{
By default this function returns a table containing all TCGA cohort
abbreviations and their corresponding disease names. A subset of that table
may be obtained by explicitly specifying one or more cohort abbreviations.
}
\examples{
format = "csv"
cohort = c("PRAD", "BRCA")

obj = Metadata.Cohorts(format = format,
                       cohort = cohort)

format = "tsv"
cohort = c("PRAD", "BRCA")

obj = Metadata.Cohorts(format = format,
                       cohort = cohort)

}

