% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.Dates.R
\name{Metadata.Dates}
\alias{Metadata.Dates}
\title{Retrieve datestamps of all GDAC Firehose standard data and analyses runs that have been ingested into FireBrowse.}
\usage{
Metadata.Dates(format = "json")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}
}
\description{
Retrieve datestamps of all GDAC Firehose standard data and analyses runs that have been ingested into FireBrowse.
}

