% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.HeartBeat.R
\name{Metadata.HeartBeat}
\alias{Metadata.HeartBeat}
\title{Simple way to discern whether API server is up and running}
\usage{
Metadata.HeartBeat(format = "json")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}
}
\description{
Returns a message to indicate that API (server) is up and running, or times out if not.
}

