% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.MAFColNames.R
\name{Metadata.MAFColNames}
\alias{Metadata.MAFColNames}
\title{Retrieve names of all columns in the mutation annotation files (MAFs) served by FireBrowse.}
\usage{
Metadata.MAFColNames(format = "json")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}
}
\description{
Retrieve the names of all columns in the mutation annotation files (MAFs) hosted by FireBrowse.  For more information on the mutation data, and how it is processed by Firehose, please consult the <a href="https://confluence.broadinstitute.org/display/GDAC/Documentation#Documentation-MutationPipelines">pipeline documentation</a>.
}

