% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.Patients.R
\name{Metadata.Patients}
\alias{Metadata.Patients}
\title{Retrieve list of all TCGA patients.}
\usage{
Metadata.Patients(format = "csv", cohort = "", page = 1,
  page_size = 250, sort_by = "cohort")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}

\item{page}{Subset to be returned.}

\item{page_size}{Number of records per page, max. is 2000.}

\item{sort_by}{character indicating the to column which is used for sorting.}
}
\description{
This service returns a list of all TCGA patient barcodes in FireBrowse,
optionally filtered by disease cohort. The barcodes are obtained directy from
the clinical data.
}
\examples{
format = "csv"
cohort  = ""
page = 1
page_size = 250
sort_by = "cohort"

obj = Metadata.Patients(format = format)

}

