% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.Platforms.R
\name{Metadata.Platforms}
\alias{Metadata.Platforms}
\title{Retrieve map of platform code(s) to platform name(s).}
\usage{
Metadata.Platforms(format = "csv", platform = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{platform}{A platforms name.}
}
\description{
By default this function returns a table of all of the technology platforms
used to sequence or characterize samples in TCGA--both their short platform
codes and full names. A subset of this table may be obtained by explicitly
specifying one or more platform codes.
}
\examples{
format = "csv"
obj = Metadata.Platforms(format = format)

format = "tsv"
obj = Metadata.Platforms(format = format)
platform = c("454", "ABI", "biotab")
obj = Metadata.Platforms(format = format,
                         platform = platform)

}

