% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.SampleType.Barcode.R
\name{Metadata.SampleType.Barcode}
\alias{Metadata.SampleType.Barcode}
\title{Given a TCGA barcode, return its short letter sample type code.}
\usage{
Metadata.SampleType.Barcode(format = "json", TCGA_Barcode = "")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{TCGA_Barcode}{Enter a single TCGA barcode, of any form: e.g. TCGA-GF-A4EO-06 or TCGA-EL-A3D5-01A-22D-A202-08}
}
\description{
Given a TCGA barcode, return its short letter sample type code.
}

