% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.SampleType.Code.R
\name{Metadata.SampleType.Code}
\alias{Metadata.SampleType.Code}
\title{Translate from numeric to symbolic TCGA sample codes.}
\usage{
Metadata.SampleType.Code(format = "json", code = "")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{code}{Narrow search to one or more TCGA sample type codes. Multiple values are allowed 01,03,07,09,12,20,40,02,04,05,06,08,10,11,13,14,50,60,61.}
}
\description{
Convert a TCGA numeric sample type code (e.g. 01, 02) to its corresponding symbolic (short letter) code (e.g. TP, TR).
}

