% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.SampleType.ShortLetterCode.R
\name{Metadata.SampleType.ShortLetterCode}
\alias{Metadata.SampleType.ShortLetterCode}
\title{Translate from symbolic to numeric TCGA sample codes.}
\usage{
Metadata.SampleType.ShortLetterCode(format = "json", short_letter_code = "")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}

\item{short_letter_code}{TCGA sample type short letter code(s) (e.g. TP, NB, etc.).  Multiple values are allowed TP,TB,TAM,TBM,NBC,CELLC,TRB,TR,TRBM,TAP,TM,THOC,NB,NT,NEBV,NBM,CELL,XP,XCL.}
}
\description{
Convert a TCGA sample type code in symbolic form (or 'short letter code' like TP, TR) to its corresponding numeric form (e.g. 01, 02).
}

