% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.SampleTypes.R
\name{Metadata.SampleTypes}
\alias{Metadata.SampleTypes}
\title{Return all TCGA sample type codes, both numeric and symbolic.}
\usage{
Metadata.SampleTypes(format = "json")
}
\arguments{
\item{format}{Format of result. Default value is json. While json,tsv,csv are available.}
}
\description{
Return all TCGA sample type codes, both numeric and symbolic.
}

