% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/External_Data_Documentation.R
\docType{data}
\name{protocols}
\alias{protocols}
\title{Indicates the protocol argument take by some functions}
\format{A matrix}
\source{
\url{http://firebrowse.org/api-docs/#!/Archives/StandardData}
}
\usage{
data(protocols)
}
\description{
Just as the \code{data_types} matrix, the protocols matrix indicates which
arguments are accepted by the two functions (function names corrospond to
column names). Also this table is hand curated and information are directly
taken from firehose Web-API.
}
\keyword{datasets}

