% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/External_Data_Documentation.R
\docType{data}
\name{sample.Type}
\alias{sample.Type}
\title{All Sample Types used by firehose}
\format{A matrix}
\source{
\url{https://tcga-data.nci.nih.gov/datareports/codeTablesReport.htm}
}
\usage{
data(sample.Type)
}
\description{
Sample Types are used by firehose to label the tissue source. Internaly the
firhose uses just abbreviation like TP or NT or even just numbers.
This matrix deciphers all those abbreviation into human readable text
}
\keyword{datasets}

