% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples.ClinicalTier1.R
\name{Samples.ClinicalTier1}
\alias{Samples.ClinicalTier1}
\title{Retrieve clinical pick tier 1 data.}
\usage{
Samples.ClinicalTier1(format = "csv", cohort = "",
  tcga_participant_barcode = "", tier1_cde_name = "", page = 1,
  page_size = 250, sort_by = "cohort")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}

\item{tcga_participant_barcode}{A character vector containing TCGA Barcodes.
Empty string returns all patients. Either one gene OR barcode is required.}

\item{tier1_cde_name}{A character vector containing the CDEs which should be
used for subsetting. To view all CDEs see
\code{\link{Metadata.ClinicalTier1}}.}

\item{page}{Subset to be returned.}

\item{page_size}{Number of records per page, max. is 2000.}

\item{sort_by}{character indicating the to column which is used for sorting.}
}
\value{
A \code{list}, if format is \code{json}, elsewise a \code{data.frame}
}
\description{
This service returns patient-level tier 1 clinical data elements (CDEs).
Results may be filtered by disease cohort, patient barcode or CDE name. When
filtering by CDE note that only when a patient record contains one or more of
the selected CDEs will it be returned. For more information on how these CDEs
are processed see our pipeline documentation \url{https://confluence.broadinstitute.org/display/GDAC/Documentation#Documentation-ClinicalPipeline}.
}
\examples{
format = "json"
cohort = ""
tcga_participant_barcode = ""
tier1_cde_name = ""
page = 1
page_size = 250
sort_by = "cohort"

obj = Samples.ClinicalTier1(format = format,
                                cohort = cohort,
                                tcga_participant_barcode = tcga_participant_barcode,
                                tier1_cde_name = tier1_cde_name,
                                page = page,
                                page_size = page_size,
                                sort_by = sort_by)

format = "csv"
obj = Samples.ClinicalTier1(format = format,
                                cohort = cohort,
                                tcga_participant_barcode = tcga_participant_barcode,
                                tier1_cde_name = tier1_cde_name,
                                page = page,
                                page_size = page_size,
                                sort_by = sort_by)

tcga_participant_barcode = "TCGA-GF-A4EO"
obj = Samples.ClinicalTier1(format = "tsv",
                                cohort = cohort,
                                tcga_participant_barcode = tcga_participant_barcode,
                                tier1_cde_name = tier1_cde_name,
                                page = page,
                                page_size = page_size,
                                sort_by = sort_by)
}

