% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/External_Data_Documentation.R
\docType{data}
\name{tools}
\alias{tools}
\title{Indicates the tools argument take by some functions}
\format{A matrix}
\source{
\url{http://firebrowse.org/api-docs/#!/Archives/StandardData}

\url{http://firebrowse.org/api-docs/#!/Archives/StandardData}
}
\usage{
data(tools)
}
\description{
Just as the \code{data_types} matrix, the tools matrix indicates which
arguments are accepted by the four functions (function names corrospond to
column names). Also this table is hand curated and informations are directly
taken from firehose Web-API.
}
\keyword{datasets}

