% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.FeatureTable.R
\name{Analyses.FeatureTable}
\alias{Analyses.FeatureTable}
\title{Retrieve aggregated analysis features table.}
\usage{
Analyses.FeatureTable(format = "tsv", cohort = "", date = "2016_01_28",
  column = "", page = "1", page_size = "250")
}
\arguments{
\item{format}{Format of result. Default value is tsv. While tsv,csv are available.}

\item{cohort}{Narrow search to one or more TCGA disease cohorts from the scrollable list. Multiple values are allowed ACC,BLCA,BRCA,CESC,CHOL,COAD,COADREAD,DLBC,ESCA,FPPP,GBM,GBMLGG,HNSC,KICH,KIPAN,KIRC,KIRP,LAML,LGG,LIHC,LUAD,LUSC,MESO,OV,PAAD,PCPG,PRAD,READ,SARC,SKCM,STAD,STES,TGCT,THCA,THYM,UCEC,UCS,UVM.}

\item{date}{Select one or more date stamps. Multiple values are allowed 2016_01_28,2015_08_21,2015_04_02,2014_10_17. Default value is 2016_01_28.}

\item{column}{Comma separated list of which data columns/fields to return. Multiple values are allowed .}

\item{page}{Which page (slice) of entire results set should be returned.  Multiple values are allowed . Default value is 1.}

\item{page_size}{Number of records per page of results.  Max is 2000. Multiple values are allowed . Default value is 250.}
}
\description{
This service returns part or all of the so-called <strong>feature table</strong>; which aggregates the most important findings across ALL pipelines in the GDAC Firehose analysis workflow into a single table for simple access.  One feature table is created per disease cohort.  Results may be filtered by date or cohort, but at least 1 cohort must be specified here. For more details please visit the <a href=https://confluence.broadinstitute.org/display/GDAC/Documentation\#Documentation-FeatureTable>online documentation</a>.  Please note that the service is still undergoing experimental evaluation and does not return JSON format.
}

