% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.mRNASeq.Quartiles.R
\name{Analyses.mRNASeq.Quartiles}
\alias{Analyses.mRNASeq.Quartiles}
\title{Returns RNASeq expression quartiles, e.g. suitable for drawing a boxplot.}
\usage{
Analyses.mRNASeq.Quartiles(format = "csv", gene = "", cohort = "",
  protocol = "RSEM", sample_type = "tumors", Exclude = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{gene}{A character vector of gene symbols. At least one gene OR Barcode
is required.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}

\item{protocol}{Should RSEM (default) or RPKM data be retrieved.}

\item{sample_type}{A character vector indicating the sample types to query.
Empty string returns all types. See \code{\link{Metadata.SampleTypes}},
for available sample types.}

\item{Exclude}{Comma separated list of TCGA participants, identified by
barcodes such as TCGA-GF-A4EO, denoting samples to exclude from computation.}
}
\value{
A \code{list}, if format is \code{json}, elsewise a \code{data.frame}
}
\description{
For a given gene compute quartiles and extrema, suitable e.g. for drawing a
boxplot (Tukey 1977). Results may be filtered by cohort, sample type, patient
barcode or characterization protocol, and are returned sorted by cohort. Note
that samples for which no expression value was recorded (e.g. the melanoma
sample TCGA-GN-262) are removed prior to calculation.
}
\examples{
format = "json"
gene = "PTEN"
cohort = ""
protocol = "RSEM"
sample_type = "tumors"
Exclude = ""

obj = Analyses.mRNASeq.Quartiles(format = format,
                                 gene = gene,
                                 cohort = cohort,
                                 protocol = protocol,
                                 sample_type = sample_type,
                                 Exclude = Exclude)

format = "tsv"
obj = Analyses.mRNASeq.Quartiles(format = format,
                                 gene = gene,
                                 cohort = cohort,
                                 protocol = protocol,
                                 sample_type = sample_type,
                                 Exclude = Exclude)
}

