% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.Centers.R
\name{Metadata.Centers}
\alias{Metadata.Centers}
\title{Retrieve map of center name(s) to display name(s).}
\usage{
Metadata.Centers(format = "csv", center = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{center}{A center name.}
}
\description{
By default this function returns a table of all consortium members in TCGA,
aka centers; it provides both the HTTP domain and full organizational name of
each center. A subset of this table may be obtained by explicitly specifying
one or more domain names.
}
\examples{
format = "csv"
center = c("intgen.org", "jhu-usc.edu")
obj = Metadata.Centers(format = format, center = center)

}

