% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.ClinicalNames_FH.R
\name{Metadata.ClinicalNames_FH}
\alias{Metadata.ClinicalNames_FH}
\title{Retrieve names of CDEs normalized by Firehose and selected for analyses}
\usage{
Metadata.ClinicalNames_FH(format = "csv")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}
}
\description{
This service returns the names of patient-level clinical data elements (CDEs)
that have been normalized by Firehose for use in analyses, unioned across all
disease cohorts. For more information on how these CDEs are processed, see
our pipeline documentation.
\url{https://confluence.broadinstitute.org/display/GDAC/Documentation}
}
\examples{
format = "csv"
obj = Metadata.ClinicalNames_FH(format = format)
}

