% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.Cohort.R
\name{Metadata.Cohort}
\alias{Metadata.Cohort}
\title{Retrieve cohort description.}
\usage{
Metadata.Cohort(format = "csv", cohort = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{cohort}{A character vector indicating the cohort to query,
empty string queries all cohorts. See \code{\link{Metadata.Cohorts}} for
available cohorts.}
}
\description{
In this client, the function does, exactly the same as
\code{\link{Metadata.Cohorts}}. It just differs within the internal
implentation. Maybe, depending on the API, this will change.
}
\examples{
format = "csv"
cohort = c("PRAD", "BRCA")

obj = Metadata.Cohort(format = format,
                       cohort = cohort)

format = "tsv"
cohort = c("PRAD", "BRCA")

obj = Metadata.Cohort(format = format,
                       cohort = cohort)
}

