% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.Dates.R
\name{Metadata.Dates}
\alias{Metadata.Dates}
\title{Retrieve list of Firehose standard data and analyses dates.}
\usage{
Metadata.Dates(format = "csv")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}
}
\description{
Retrieve list of Firehose standard data and analyses dates.
}
\examples{
format = "csv"
obj = Metadata.Dates(format = format)

format = "tsv"
obj = Metadata.Dates(format = format)
}

