% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.HeartBeat.R
\name{Metadata.HeartBeat}
\alias{Metadata.HeartBeat}
\title{Simple way to discern whether API server is up and running}
\usage{
Metadata.HeartBeat(format = "csv")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}
}
\description{
Returns a message to indicate that API (server) is up and running, or times
out if not.
}
\examples{
format = "csv"
obj = Metadata.HeartBeat(format = format)
}

