% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Metadata.Platforms.R
\name{Metadata.Platforms}
\alias{Metadata.Platforms}
\title{Retrieve map of platform code(s) to platform name(s).}
\usage{
Metadata.Platforms(format = "csv", platform = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{platform}{A platforms name.}
}
\description{
To return a map of all available platform codes to platform names, exclude
the platform parameter. Otherwise, choose only the platforms you wish to be
included in the result.
}
\examples{
format = "csv"
obj = Metadata.Platforms(format = format)

format = "tsv"
obj = Metadata.Platforms(format = format)
platform = c("454", "ABI", "biotab")
obj = Metadata.Platforms(format = format,
                         platform = platform)
}

