% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Metadata.SampleType.Code.R
\name{Metadata.SampleType.Code}
\alias{Metadata.SampleType.Code}
\title{Translate from numeric sample type code to short letter code.}
\usage{
Metadata.SampleType.Code(format = "csv", code = "")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}

\item{code}{Sample type codes.}
}
\description{
Translate a numeric sample type code (e.g. 01, 02, etc.) into its short
letter code (e.g. TP, TR, etc.).
}
\examples{
format = "csv"
code = c(50, 60)
obj = Metadata.SampleType.Code(format = format, code = code)
}

