% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.SampleTypes.R
\name{Metadata.SampleTypes}
\alias{Metadata.SampleTypes}
\title{Return a sample type code to short letter code mapping.}
\usage{
Metadata.SampleTypes(format = "csv")
}
\arguments{
\item{format}{A string identifying the data type returned. Using json
requires the jsonlite package. \code{json}, \code{csv} or \code{tsv} are
available. In this package \code{tsv} and \code{csv} are identical.}
}
\description{
Return a sample type code to short letter code mapping.
}
\examples{
format = "csv"
obj = Metadata.SampleTypes(format = format)

format = "tsv"
obj = Metadata.SampleTypes(format = format)
}

